@@ -1,21 +1,21 @@
 name: upgrader
 description: Flutter package for prompting users to upgrade when there is a newer version of the app in the store.
-version: 6.0.1
+version: 7.0.0
 homepage: https://github.com/larryaasen/upgrader
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
-  flutter: ">=3.3.10"
+  sdk: '>=2.18.6 <4.0.0'
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # From fluttercommunity.dev: Get current device information from within the Flutter application.
-  device_info_plus: ">=3.2.0 <8.0.0"
+  device_info_plus: ^9.0.0
 
   # Pure Dart library for HTML5 parsing
-  html: ^0.15.1
+  html: ^0.15.3
 
   # For app_theme.dart
   google_fonts: ^2.3.1
@@ -24,32 +24,32 @@ dependencies:
   google_api_availability: ^3.0.1
 
   # From Dart Team: A composable, Future-based library for making HTTP requests.
-  http: ^0.13.5
+  http: ^0.13.6
 
   # From dart.dev: Platform independent access to information about the current operating system.
   os_detect: ^2.0.1
 
   # From Flutter Community: Provides an API for querying information about an application package.
-  package_info_plus: '>=1.3.0 <3.0.3'
+  package_info_plus: ^4.0.0
 
   # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
-  shared_preferences: ^2.0.15
+  shared_preferences: ^2.1.1
 
   # From Flutter Team: A Flutter plugin for launching a URL in the mobile platform.
-  url_launcher: ^6.1.6
+  url_launcher: ^6.1.11
 
   # A dart library for comparing and incrementing version numbers with Semantic Versioning.
   version: ^3.0.2
 
   # A dart library for parsing, traversing, querying, transforming and building XML documents.
-  xml: ^6.1.0
+  xml: ^6.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   # From Dart Team: Mock library for Dart inspired by Mockito.
-  mockito: ^5.3.2
+  mockito: ^5.4.0
   flutter_lints: ^2.0.1
 
 flutter:
