@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2019-2023 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/material.dart';
@@ -11,7 +11,8 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, setup the Appcast.
+  // On Android, the default behavior will be to use the Google Play Store
+  // version of the app.
   // On iOS, the default behavior will be to use the App Store version of
   // the app, so update the Bundle Identifier in example/ios/Runner with a
   // valid identifier already in the App Store.
@@ -19,27 +20,21 @@ void main() async {
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final appcastURL =
-        'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
-    final cfg = AppcastConfiguration(url: appcastURL, supportedOS: ['android']);
-
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-          appBar: AppBar(title: Text('Upgrader Example')),
-          body: Center(
-              child: Container(
-                  margin: EdgeInsets.fromLTRB(12.0, 0.0, 12.0, 0.0),
-                  child: UpgradeCard(
-                    upgrader: Upgrader(
-                      appcastConfig: cfg,
-                      debugLogging: true,
-                    ),
-                  )))),
+        appBar: AppBar(title: Text('Upgrader Example')),
+        body: Center(
+          child: Container(
+            margin: EdgeInsets.fromLTRB(12.0, 0.0, 12.0, 0.0),
+            child: UpgradeCard(),
+          ),
+        ),
+      ),
     );
   }
 }
