@@ -6,12 +6,17 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
 import 'package:upgrader/upgrader.dart';
 
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  setMockDeviceInfo();
+
   setUp(() async {});
 
   tearDown(() async {});
@@ -49,6 +54,27 @@ void main() {
     final items = await appcast.parseAppcastItemsFromUri(
         'https://sparkle-project.org/test/testappcast.xml');
     validateItems(items!, appcast);
+  }, skip: false);
+
+  test(
+      'testing Appcast will prioritize critical version even with lover version. ',
+      () async {
+    final appcast = TestAppcast();
+
+    final testFile =
+        await getTestFile(filePath: 'test/testappcast_critical.xml');
+
+    await appcast.parseAppcastItemsFromFile(testFile);
+
+    final bestCriticalItem = appcast.bestCriticalItem();
+
+    expect(
+      bestCriticalItem?.versionString == "3.0.0",
+      equals(true),
+    );
+
+    expect(bestCriticalItem?.tags?.contains("sparkle:criticalUpdate"),
+        equals(true));
   }, skip: false);
 
   test('testing Appcast host', () async {
@@ -132,8 +158,8 @@ void validateItems(List<AppcastItem> items, Appcast appcast) {
   expect(bestItem.osString, isNull);
 }
 
-Future<File> getTestFile() async {
-  var testFile = File('test/testappcast.xml');
+Future<File> getTestFile({String filePath = 'test/testappcast.xml'}) async {
+  var testFile = File(filePath);
   final exists = await testFile.exists();
   if (!exists) {
     testFile = File('testappcast.xml');
@@ -141,14 +167,15 @@ Future<File> getTestFile() async {
   return testFile;
 }
 
-Future<http.Client> setupMockClient() async {
+Future<http.Client> setupMockClient(
+    {String filePath = 'test/testappcast.xml'}) async {
   // Use a mock to return a successful response when it calls the
   // provided http.Client.
 
   final client = MockClient((http.Request request) async {
     if (request.url.toString() ==
         'https://sparkle-project.org/test/testappcast.xml') {
-      final testFile = await getTestFile();
+      final testFile = await getTestFile(filePath: filePath);
       final contents = await testFile.readAsString();
       return http.Response.bytes(utf8.encode(contents), 200);
     }
@@ -164,4 +191,39 @@ class TestAppcast extends Appcast {
     final contents = await file.readAsString();
     return parseAppcastItems(contents);
   }
+}
+
+void setMockDeviceInfo() {
+  const channel = MethodChannel('dev.fluttercommunity.plus/device_info');
+
+  handler(MethodCall methodCall) async {
+    print('setMockDeviceInfo.methodCall: ${methodCall.method}');
+
+    switch (methodCall.method) {
+      case 'getMacosDeviceInfo':
+      case 'getDeviceInfo':
+        return <String, dynamic>{
+          'computerName': '',
+          'hostName': '',
+          'arch': '',
+          'model': '',
+          'kernelVersion': '',
+          'osRelease': 'Version 13.2.1 (Build 22D68)',
+          'majorVersion': 13,
+          'minorVersion': 2,
+          'patchVersion': 1,
+          'activeCPUs': 0,
+          'memorySize': 0,
+          'cpuFrequency': 0,
+          'systemGUID': '',
+        };
+
+      default:
+        assert(false);
+        return null;
+    }
+  }
+
+  TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+      .setMockMethodCallHandler(channel, handler);
 }
